Fetch is now sync command; improved behavior
The fetch subcommand has been renamed to sync.  This is to avoid
confusion with the `git fetch` behavior.

Behavior of sync has changed so that for existing repos, if the worktree
is unclean a fetch is performed, otherwise a pull is attempted.

Tests are built around the sync functionality.

Both sync and parse now take a channel of errors.  This is used to send
errors that occur that do not need to stop the process.  For example, if
a repo cannot sync the error is sent through the channel but we continue
to sync the next repo.